wip: Quick prototype of reorganizing poller+manager to better separate concerns
